pyproject.toml
[project]
name = "pdf2imgCrop"
version = "0.1.1"
description = "将PDF文件转换为图片并自动裁剪空白边距"
readme = "README.md"
authors = [
    {name = "Muxkin", email = "your.email@example.com"},
]
dependencies = [
    "PyMuPDF>=1.23.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
]
requires-python = ">=3.8"
keywords = ["pdf", "image", "convert", "crop"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]

[project.urls]
"Bug Tracker" = "https://github.com/muxkin/pdf2imgCrop/issues"
"Source Code" = "https://github.com/muxkin/pdf2imgCrop"

[project.scripts]
pdf2imgcrop = "pdf2imgCrop.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pdf2imgCrop"]
